style(components) : rename file
