+++ src/components/curaded-artwork/CuratedArtworkSection.tsx
